+++ MSFSTouchPortalPlugin/Services/PluginService.cs
@@ -71,6 +71,8 @@ namespace MSFSTouchPortalPlugin.Services {
             // Handle conversions
             if (ShouldConvertToFload(value.Unit)) {
               valObj = float.Parse(stringVal);
+            } else if (value.Unit == Units.String) {
+              valObj = ((StringVal64)data).Value;
             } else if (value.Unit == Units.radians) {
               // Convert to Degrees
               valObj = float.Parse(stringVal) * (180 / Math.PI);